feat: Master key -> keychain key -> keychain
Having the keychain key directly derived from the
master password is risky as there is no way to rotate
the credentials and keychain independently.

Having the master key used only to unlock the
keychain key makes it easier to store the keychain
key locally during the session, and rotating it when
necessary.